refactor(trading-signals): rename BandsResult.ts to BandsResult.types.ts
Applies the .types.ts suffix for pure type files and converts the
interface to a closed type alias — a result shape is a fixed data
contract, not something consumers should merge into or extend.
