+++ packages/trading-signals/src/base/index.ts
@@ -1,4 +1,4 @@
-export * from './BandsResult.js';
+export * from './BandsResult.types.js';
 export * from './Candle.types.js';
 export * from './Indicator.js';
 export * from './Period.js';
