+++ packages/trading-signals/src/volatility/ABANDS/AccelerationBands.ts
@@ -2,7 +2,7 @@ import {TechnicalIndicator, TradingSignal} from '../../base/Indicator.js';
 import type {MovingAverage} from '../../trend/MA/MovingAverage.js';
 import type {MovingAverageTypes} from '../../trend/MA/MovingAverageTypes.js';
 import {SMA} from '../../trend/SMA/SMA.js';
-import type {BandsResult} from '../../base/BandsResult.js';
+import type {BandsResult} from '../../base/BandsResult.types.js';
 import type {HighLowClose} from '../../base/Candle.types.js';
 
 /**
